separate stub class in specific file
