+++ src/test/ut/java/fr/cnes/sonar/report/CommonTest.java
@@ -48,30 +48,11 @@ public abstract class CommonTest {
     protected ReportConfiguration conf;
 
     /**
-     * Stub class to return non-null rules without web call
-     */
-    class stubReport extends Report{
-        @Override
-        public Rule getRule(String key){
-            Rule rule = new Rule();
-            rule.setSeverity(MAJOR);
-            rule.setHtmlDesc("<p>description</p>");
-            rule.setKey("c");
-            rule.setName("squid:1234");
-            rule.setType("BUG");
-            rule.setDebt("2h");
-            rule.setLang("FR");
-            rule.setLangName("Francais");
-            return rule;
-        }
-    }
-
-    /**
      * Setting of all stubbed resources before launching a test.
      */
     @Before
     public void before() {
-        report = new stubReport();
+        report = new StubReport();
         conf = ReportConfiguration.create(new String[]{
                 "-s", "http://sonarqube:9000",
                 "-p", "cnesreport",
